Oauth : google oauth login api done
/auth/google
fail => redirect(/login)
success => res.json(req.user)

+++ server/models/booking.js
@@ -2,16 +2,16 @@
 module.exports = (sequelize, DataTypes) => {
   const Booking = sequelize.define('booking', {
     booking_id: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.BIGINT(11),
         primaryKey: true
     },
     check_in : DataTypes.STRING,
     check_out: DataTypes.STRING,
-    fk_guest_id: DataTypes.INTEGER,
+    fk_guest_id: DataTypes.BIGINT(11),
     guest_name: DataTypes.STRING,
-    host_id: DataTypes.INTEGER,
+    host_id: DataTypes.BIGINT(11),
     host_name: DataTypes.STRING,
-    fk_room_id: DataTypes.INTEGER,
+    fk_room_id: DataTypes.BIGINT(11),
     room_name: DataTypes.STRING,
     guest_total: DataTypes.INTEGER,
     guest_adult: DataTypes.INTEGER,
