+++ server/models/review.js
@@ -2,11 +2,11 @@
 module.exports = (sequelize, DataTypes) => {
   const Review = sequelize.define('review', {
     review_id: {
-        type:DataTypes.INTEGER,
+        type:DataTypes.BIGINT(11),
         primaryKey: true
     },
-    fk_room_id: DataTypes.INTEGER,
-    guest_id: DataTypes.INTEGER,
+    fk_room_id: DataTypes.BIGINT(11),
+    guest_id: DataTypes.BIGINT(11),
     guest_name: DataTypes.STRING,
     host_name: DataTypes.STRING,
     comment: DataTypes.STRING,
