+++ server/models/user.js
@@ -2,7 +2,7 @@
 module.exports = (sequelize, DataTypes) => {
   const User = sequelize.define('user', {
     user_id: {
-        type:DataTypes.INTEGER,
+        type:DataTypes.BIGINT(11),
         primaryKey: true
     },
     name: DataTypes.STRING,
